Convert HeartFrequency from Float to Double

+++ app/src/main/java/com/inky/fitnesscalendar/data/measure/HeartFrequency.kt
@@ -7,14 +7,14 @@ import kotlin.math.roundToInt
 
 @Serializable
 @JvmInline
-value class HeartFrequency internal constructor(val bpm: Float) : Comparable<HeartFrequency>,
+value class HeartFrequency internal constructor(val bpm: Double) : Comparable<HeartFrequency>,
     Measure {
     override fun compareTo(other: HeartFrequency) = this.bpm.compareTo(other.bpm)
 
-    override fun isNothing() = bpm == 0f
+    override fun isNothing() = bpm == 0.0
 
     override fun formatWithContext(context: Context) =
         context.getString(R.string.x_bpm, bpm.roundToInt())
 }
 
-fun Float.bpm() = HeartFrequency(bpm = this)
+fun Double.bpm() = HeartFrequency(bpm = this)
